Bank Customer Churn/requirements.txt
lightgbm
pandas
seaborn
matplotlib
scikit-learn
xgboost
jupyter
joblib
numpy
pydantic
sqlalchemy
uvicorn
fastapi
imblearn
streamlit
python-dotenv
psycopg2-binary
python-multipart
requests
great-expectations
Unidecode